Remove unnecessary file path comments.
