+++ shared/seismic/envelopes/envelope-config.ts
@@ -1,5 +1,3 @@
-// shared/seismic/envelope-config.ts
-
 /**
  * Envelope tile cache level configuration.
  *
