+++ shared/seismic/envelopes/tile-addressing.test.ts
@@ -1,6 +1,6 @@
 import {
   getTileIndex, getTileTimeRange, getTileIndicesForViewport, getTileS3Key, getTileDuration, getPointIndexInTile,
-  getStationPrefix, parseStationPrefix, encodeLocation, decodeLocation, getS3Root
+  getS3Root
 } from "./tile-addressing";
 import { LEVEL_SPACINGS, POINTS_PER_TILE } from "./envelope-config";
 
@@ -76,21 +76,6 @@ describe("tile-addressing", () => {
     });
   });
 
-  describe("encodeLocation / decodeLocation", () => {
-    it("encodes blank locations as '--'", () => {
-      expect(encodeLocation("")).toBe("--");
-      expect(encodeLocation(undefined)).toBe("--");
-      expect(encodeLocation("00")).toBe("00");
-    });
-
-    it("round-trips through decodeLocation", () => {
-      expect(decodeLocation(encodeLocation(""))).toBe("");
-      expect(decodeLocation(encodeLocation("00"))).toBe("00");
-      // A literal "--" is the path encoding of blank, so it normalizes to "" rather than surviving.
-      expect(decodeLocation(encodeLocation("--"))).toBe("");
-    });
-  });
-
   describe("getS3Root", () => {
     it("appends the current layout version", () => {
       expect(getS3Root("base/")).toBe("base/v2/");
@@ -106,20 +91,6 @@ describe("tile-addressing", () => {
     it("includes a non-blank location code", () => {
       const key = getTileS3Key({ network: "IU", station: "ANMO", location: "00", channel: "BHZ" }, 2, 42);
       expect(key).toBe("IU_ANMO/00/BHZ/L2/42");
-    });
-  });
-
-  describe("parseStationPrefix", () => {
-    it("is the inverse of getStationPrefix", () => {
-      expect(parseStationPrefix("AK_K204")).toEqual({ network: "AK", station: "K204" });
-      const s = { network: "AK", station: "RC01" };
-      expect(parseStationPrefix(getStationPrefix(s))).toEqual(s);
-    });
-
-    it("returns undefined for improper prefixes", () => {
-      expect(parseStationPrefix("prefix")).toBeUndefined();
-      expect(parseStationPrefix("_prefix")).toBeUndefined();
-      expect(parseStationPrefix("prefix_")).toBeUndefined();
     });
   });
 });
