+++ shared/seismic/models/seismic-architectures.test.ts
