+++ shared/seismic/models/seismic-architectures.ts
