+++ shared/seismic/models/seismic-model-types.ts
