+++ src/models/stores/seismic/seismic-coverage-processor.test.ts
@@ -1,6 +1,6 @@
 import { dayRange, SECONDS_PER_DAY } from "../../../../shared/seismic/seismic-day";
-import { SeismicModelRunner } from "../../../../shared/seismic/seismic-model-runner";
-import { ModelMetadata } from "../../../../shared/seismic/seismic-model-types";
+import { SeismicModelRunner } from "../../../../shared/seismic/models/seismic-model-runner";
+import { ModelMetadata } from "../../../../shared/seismic/models/seismic-model-types";
 import { StationData, TimeRange } from "../../../../shared/seismic/seismic-types";
 import { DETECTION_THRESHOLD, processUncoveredRanges } from "./seismic-coverage-processor";
 import { FakeDownloadService, makeFakeDownloadService, makeFakeModelRunner } from "./seismic-coverage-test-fakes";
