+++ src/models/stores/seismic/seismic-query-service.test.ts
@@ -1,8 +1,8 @@
 import { DateTime } from "luxon";
 import { SeismicQueryService, envelopeCacheKey } from "./seismic-query-service";
-import { encodeEnvelopeTile } from "../../../../shared/seismic/envelope-codec";
-import { LEVEL_SPACINGS, NO_DATA_SENTINEL } from "../../../../shared/seismic/envelope-config";
-import { getTileTimeRange } from "../../../../shared/seismic/tile-addressing";
+import { encodeEnvelopeTile } from "../../../../shared/seismic/envelopes/envelope-codec";
+import { LEVEL_SPACINGS, NO_DATA_SENTINEL } from "../../../../shared/seismic/envelopes/envelope-config";
+import { getTileTimeRange } from "../../../../shared/seismic/envelopes/tile-addressing";
 import { SeismicViewportParams } from "../../../../shared/seismic/seismic-types";
 
 // Mock fetch globally
