+++ src/plugins/wave-runner/components/data-setup.tsx
@@ -1,6 +1,6 @@
 import { observer } from "mobx-react";
 import React, { useEffect, useMemo } from "react";
-import { ModelListEntry } from "../../../../shared/seismic/model-metadata";
+import { ModelListEntry } from "../../../../shared/seismic/models/model-metadata";
 import { StationConfig } from "../../../../shared/seismic/seismic-types";
 import { useSettingFromStores } from "../../../hooks/use-stores";
 import { stationId } from "../../shared-seismogram/station-model";
