+++ src/seismic-admin/components/admin-body.tsx
@@ -1,6 +1,6 @@
 import { observer } from "mobx-react-lite";
 import React from "react";
-import { getStationChannelPrefix } from "../../../shared/seismic/tile-addressing";
+import { getStationChannelPrefix } from "../../../shared/seismic/station-addressing";
 import { useSeismicAdminStore } from "../hooks/use-seismic-admin-stores";
 import { StationSection } from "./station-section";
